Fixing standard issues
